move rate rule name in entry.data.comment
the rate rule name in entry.data.comment was at the start of
the comment line, which caused issues in reading chemkin comments.
It was moved to the end of the comment to allow parsing the start
of chemkin lines.

Also, with the new description, the parsing for uncertainty needs
to be changed so it can catch the different syntax for training
reactions.